refactor: 수업 교환 페이지 컴포넌트 분리
