+++ components/staff/class/ExchangeRequestActionBar.tsx
@@ -1,0 +1,67 @@
+"use client";
+
+import styled from "styled-components";
+import { Button } from "@/components/common/VariantButton";
+import { layout, spacing } from "@/styles/tokens";
+
+interface ExchangeRequestActionBarProps {
+  canDelete: boolean;
+  canEdit: boolean;
+  isEditing: boolean;
+  isUpdating: boolean;
+  onDelete: () => void;
+  onCancelEdit: () => void;
+  onSaveEdit: () => void;
+  onStartEdit: () => void;
+}
+
+export function ExchangeRequestActionBar({
+  canDelete,
+  canEdit,
+  isEditing,
+  isUpdating,
+  onDelete,
+  onCancelEdit,
+  onSaveEdit,
+  onStartEdit,
+}: ExchangeRequestActionBarProps) {
+  return (
+    <TopButtonRow>
+      <Button type="button" $variant="danger" disabled={!canDelete} onClick={onDelete}>
+        삭제
+      </Button>
+
+      {isEditing ? (
+        <>
+          <Button type="button" $variant="neutral" onClick={onCancelEdit} disabled={isUpdating}>
+            취소
+          </Button>
+
+          <Button type="button" onClick={onSaveEdit} disabled={isUpdating}>
+            저장
+          </Button>
+        </>
+      ) : (
+        <Button type="button" onClick={onStartEdit} disabled={!canEdit}>
+          수정
+        </Button>
+      )}
+    </TopButtonRow>
+  );
+}
+
+const TopButtonRow = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  gap: ${spacing.space20};
+  margin-bottom: 2.25rem;
+
+  @media (min-width: 120rem) {
+    gap: 1.875rem;
+    margin-bottom: 3rem;
+  }
+
+  @media (max-width: ${layout.breakpointMobile}) {
+    flex-wrap: wrap;
+  }
+`;
